% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testnestedclang-package.R
\docType{package}
\name{testnestedclang-package}
\alias{testnestedclang}
\alias{testnestedclang-package}
\title{testnestedclang: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
